update: address comment
